pyproject.toml
[project]
name = "codinglegends"
version = "0.1.0"
description = "This project aims at automating ICD Coding using an agentic approach. It builds on top of the codeseeker project provided by the paper Code Like Humans"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "datasets<3.0.0",
    "intervaltree>=3.1.0",
    "ipykernel>=7.1.0",
    "jinja2>=3.1.0",
    "jupyter>=1.1.1",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "loguru>=0.7.0",
    "langsmith>=0.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "polars>=0.20.0,<1.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.7",
    "pyyaml>=6.0",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "seaborn>=0.13.2",
    "spacy>=3.7.0",
    "tqdm>=4.67.1",
]
